Move CopyTextArea and organize stories index.
Moved CopyTextArea and its stories to a new util package.
