+++ react-components/src/apps/details.js
@@ -1,9 +1,7 @@
 import CategoryTree from './details/CategoryTree';
-import CopyTextArea from './details/CopyTextArea';
 import ToolDetails from './details/ToolDetails';
 
 export {
     CategoryTree,
-    CopyTextArea,
     ToolDetails,
 };
