+++ react-components/src/util/CopyTextArea.js
@@ -4,7 +4,7 @@
 import React, { Component } from 'react';
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
-import {hasClipboardAPI, copySelection} from '../../clipboardFunctions'
+import {hasClipboardAPI, copySelection} from '../clipboardFunctions'
 
 class CopyTextArea extends Component {
     constructor(props) {
