+++ react-components/stories/index.stories.js
@@ -3,10 +3,10 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import CopyTextAreaTest from './apps/details/CopyTextArea.stories'
 import ToolDetailsTest from './apps/details/ToolDetails.stories'
 import CategoryTreeTest from './apps/details/CategoryTree.stories'
+import CopyTextAreaTest from './util/CopyTextArea.stories'
 
-storiesOf('CopyTextArea', module).add('with test text', () => <CopyTextAreaTest/>);
-storiesOf('CategoryTree', module).add('with test hierarchy', () => <CategoryTreeTest logger={action('hierarchy')} />);
-storiesOf('ToolDetails', module).add('with test app', () => <ToolDetailsTest/>);
+storiesOf('apps/details', module).add('CategoryTree', () => <CategoryTreeTest logger={action('hierarchy')} />);
+storiesOf('apps/details', module).add('ToolDetails', () => <ToolDetailsTest/>);
+storiesOf('util', module).add('CopyTextArea', () => <CopyTextAreaTest/>);
